refactor lol
