+++ package.json
@@ -1,0 +1,5 @@
+{
+  "devDependencies": {
+    "vite": "^5.4.10"
+  }
+}
